Move Model/BaseModel to new package
Will help avoid future import cycles.
