+++ src/internal/kopia/model_store.go
@@ -7,6 +7,8 @@ import (
 	"github.com/kopia/kopia/repo"
 	"github.com/kopia/kopia/repo/manifest"
 	"github.com/pkg/errors"
+
+	"github.com/alcionai/corso/internal/model"
 )
 
 const stableIDKey = "stableID"
@@ -25,46 +27,6 @@ const (
 	RestoreOpModel
 	RestorePointModel
 )
-
-type ID string
-
-type Model interface {
-	GetStableID() ID
-	SetStableID(id ID)
-	GetModelStoreID() manifest.ID
-	SetModelStoreID(id manifest.ID)
-}
-
-// BaseModel defines required fields for models stored in ModelStore. Structs
-// that wish to be stored should embed this struct.
-type BaseModel struct {
-	// StableID is an identifier that other objects can use to refer to this
-	// object in the ModelStore.
-	// Once generated (during Put), it is guaranteed not to change. This field
-	// should be treated as read-only by users.
-	StableID ID `json:"stableID,omitempty"`
-	// ModelStoreID is an internal ID for the model in the store. If present it
-	// can be used for efficient lookups, but should not be used by other models
-	// to refer to this one. This field may change if the model is updated. This
-	// field should be treated as read-only by users.
-	ModelStoreID manifest.ID `json:"modelStoreID,omitempty"`
-}
-
-func (bm *BaseModel) GetStableID() ID {
-	return bm.StableID
-}
-
-func (bm *BaseModel) SetStableID(id ID) {
-	bm.StableID = id
-}
-
-func (bm *BaseModel) GetModelStoreID() manifest.ID {
-	return bm.ModelStoreID
-}
-
-func (bm *BaseModel) SetModelStoreID(id manifest.ID) {
-	bm.ModelStoreID = id
-}
 
 func NewModelStore(kw *KopiaWrapper) *ModelStore {
 	return &ModelStore{wrapper: kw}
@@ -101,7 +63,7 @@ func tagsForModel(t modelType, tags map[string]string) (map[string]string, error
 // type or if a bad model type is given.
 func tagsForModelWithID(
 	t modelType,
-	id ID,
+	id model.ID,
 	tags map[string]string,
 ) (map[string]string, error) {
 	if len(id) == 0 {
@@ -129,13 +91,13 @@ func putInner(
 	w repo.RepositoryWriter,
 	t modelType,
 	tags map[string]string,
-	m Model,
+	m model.Model,
 	create bool,
 ) error {
 	// ModelStoreID does not need to be persisted in the model itself.
 	m.SetModelStoreID("")
 	if create {
-		m.SetStableID(ID(uuid.NewString()))
+		m.SetStableID(model.ID(uuid.NewString()))
 	}
 
 	tmpTags, err := tagsForModelWithID(t, m.GetStableID(), tags)
@@ -160,7 +122,7 @@ func (ms *ModelStore) Put(
 	ctx context.Context,
 	t modelType,
 	tags map[string]string,
-	m Model,
+	m model.Model,
 ) error {
 	err := repo.WriteSession(
 		ctx,
@@ -186,19 +148,23 @@ func (ms *ModelStore) GetIDsForType(
 	ctx context.Context,
 	t modelType,
 	tags map[string]string,
-) ([]ID, error) {
+) ([]model.ID, error) {
 	return nil, nil
 }
 
 // Get deserializes the model with the given ID into data.
-func (ms *ModelStore) Get(ctx context.Context, id ID, data any) error {
+func (ms *ModelStore) Get(ctx context.Context, id model.ID, data any) error {
 	return nil
 }
 
 // GetWithModelStoreID deserializes the model with the given ModelStoreID into
 // data. Returns github.com/kopia/kopia/repo/manifest.ErrNotFound if no model
 // was found.
-func (ms *ModelStore) GetWithModelStoreID(ctx context.Context, id manifest.ID, data Model) error {
+func (ms *ModelStore) GetWithModelStoreID(
+	ctx context.Context,
+	id manifest.ID,
+	data model.Model,
+) error {
 	if len(id) == 0 {
 		return errors.WithStack(errNoModelStoreID)
 	}
@@ -220,14 +186,14 @@ func (ms *ModelStore) GetWithModelStoreID(ctx context.Context, id manifest.ID, d
 func (ms *ModelStore) Update(
 	ctx context.Context,
 	t modelType,
-	oldID ID,
+	oldID model.ID,
 	tags map[string]string,
 	m any,
-) (ID, error) {
+) (model.ID, error) {
 	return "", nil
 }
 
 // Delete deletes the model with the given ID from the model store.
-func (ms *ModelStore) Delete(ctx context.Context, id ID) error {
+func (ms *ModelStore) Delete(ctx context.Context, id model.ID) error {
 	return nil
 }
